+++ src/components/form/MultifieldGroup/multifieldGroup.stories.js
@@ -1,5 +1,6 @@
 import multifieldGroup from './multifieldGroup.html.twig';
 import multifieldGroupContent from './multifieldGroupContent.html.twig';
+import multifieldGroupContentFirstStep from './multifieldGroupContentFirstStep.html.twig';
 import './multifieldGroup.scss';
 
 export default {
@@ -7,7 +8,8 @@ export default {
 };
 
 const Template = (args) => multifieldGroup(args);
-const ContentTemplate = (args) => multifieldGroupContent({ ...args, step_number: 1 });
+const ContentTemplate = (args) => multifieldGroupContent({ ...args });
+const FirstItemContentTemplate = (args) => multifieldGroupContentFirstStep({ ...args });
 
 export const Default = Template.bind({});
 Default.args = {
@@ -75,6 +77,46 @@ FrozenSteps.args = {
         allergen: [],
       },
       errors: ['step_name'],
+    },
+  ],
+  label: 'Enter the process step details',
+  entity_name: 'Step',
+  remove_this: 'Remove this',
+  add_another: 'Add another',
+};
+
+export const AnotherTemplateForFirstItem = Template.bind({});
+AnotherTemplateForFirstItem.args = {
+  template: ContentTemplate.bind({}),
+  first_item_template: FirstItemContentTemplate.bind({}),
+  id: 'multi-steps',
+  required: true,
+  items: [
+    {
+      values: {
+        step_number: '3',
+        step_name: 'Cooking chicken',
+        biological: ['salmonella'],
+        chemical: ['Bleach', 'Another chemical'],
+        physical: ['Hair'],
+        allergen: [],
+        delivery_method: 'sms',
+        notes: 'nothing',
+      },
+      errors: ['chemical', 'allergen'],
+    },
+    {
+      values: {
+        step_number: '5',
+        step_name: 'Cooking fish',
+        biological: ['salmonella'],
+        chemical: ['Bleach', 'Another chemical'],
+        physical: ['Hair'],
+        allergen: [],
+        notes: '',
+      },
+      errors: ['step_name'],
+      delivery_method: 'email',
     },
   ],
   label: 'Enter the process step details',